Show organization events on the home page
The organization (account) events are fetched from the server for
each account and merged to a full list, sorted by creation time,
thus giving a timeline of all the changes in all the organizations
the user has access to.

The events are cached to session and there is an option to refresh
them by a click of a button.

It would make sense to fetch also the other events and merge them
to this one, making a real full event stream that shows all events
that have happened in all the organizations and then add a filtering
option to the client side. These could be also cached to disk since
these are actually immutable so these should be safe to cache forever.